Version 1.0.0 (#1)
* New. Common. Bot detector including implemented.

* New. Common. Event token support implemented.

* Fix. Version fixed.

* Fix. Code. Front modules removed.

* Upd. Code. Classes structure updated.

* Fix. Code. SFW disabled temporary.

* Fix. Settings. Settings saving fixed.

* Fix. Settings. Option `ct_posts_to_check` moved to the up of the settings page.

* Fix. Application. Method `spamCheck()` modified.

* New. Anti-Spam. New posts protection implemented.

* New. Application. Option `Show link to CleanTalk` implemented.

* Fix. Anti-Spam. New posts protection fixed.

* Fix. Anti-Spam. Register protection fixed.

* New. Anti-Spam. ContactUs protection implemented.

* Fix. Settings. Spam users checking fixed.

* Fix. Admin notice. Error on admin notice fixed.

* Fix. Register. Exclude checking of register users by admin.

* Fix. Register. Block massage implemented.

* Fix. Install. Install error fixed.

* Fix. Users checker. Removed functionality for implement this in the future.

* Fix. SFW. SFW running disabled temporary.

* Fix. RC. Remote Calls running disabled temporary.

* Upd. Rebuild version 1.0.0.

+++ listeners/Register.php
@@ -29,13 +29,28 @@ if ( !defined('\IPS\SUITE_UNIQUE_KEY') ) {
  */
 class Register extends MemberListenerType
 {
-    public function onCreateAccount(MemberClass $member)
+    public function onCreateAccount(MemberClass $member): void
     {
-        $ct_result = Application::spamCheck(true);
+        $current_user = \IPS\Member::loggedIn();
+        if ( $current_user && $current_user->isAdmin() ) {
+            error_log(var_export($member->get_name() . ' register check excluded.', 1));
+            return;
+        }
+
+        $request_params = [
+            'post_info' => [
+                'comment_type' => 'register'
+            ],
+            'sender_email' => $member->email,
+            'sender_nickname' => $member->get_name(),
+        ];
+
+        $ct_result = Application::spamCheck($request_params, true);
 
         if ( $ct_result && isset($ct_result->errno) && $ct_result->errno == 0 && $ct_result->allow == 0 ) {
             // Spammer - delete this user
             $member->delete();
+            die($ct_result->comment);
         }
     }
 }
